+++ IGraphics/IGraphicsPathBase.h
@@ -556,7 +556,7 @@ protected:
     double tx, ty;
     
     CalulateTextRotation(text, bounds, rotated, tx, ty);
-    PathTransformTranslate(tx, ty);
+    PathTransformTranslate(static_cast<float>(tx), static_cast<float>(ty));
     PathTransformRotate(text.mAngle);
   }
